tests/users/test_forms.py
from django.urls import reverse
import pytest
from app.users import forms
from app.users.models import Users


@pytest.mark.django_db
def test_user_registration_form_succsess():
    form = forms.UserRegistrationForm(
        data={
            'username': 'test3@test.com',
            'password1': 'testpassword',
            'password2': 'testpassword',
        },
    )
    assert form.is_valid() is True
    data = form.save()
    assert data['next_url'] == reverse('login_template')
    user = Users.objects.get(email='test3@test.com')
    assert user.email == 'test3@test.com'


@pytest.mark.django_db
def test_user_registration_form_fail():
    form = forms.UserRegistrationForm(
        data={
            'username': 'test3@test.com',
            'password1': 'testpassword',
            'password2': 'testpassword2',
        },
    )
    assert form.is_valid() is False
